requirements.txt
customtkinter>=5.2.0
mysql-connector-python>=8.0.33
bcrypt>=4.0.1
pandas>=2.0.0
matplotlib>=3.7.0
reportlab>=4.0.0   # For PDF export (settings tab)
sqlalchemy>=2.0.0  # Required for pandas read_sql compatibility
pymysql>=1.0.0     # SQLAlchemy MySQL dialect driver
